requirements.txt
pytest
pytest-pycodestyle
requests
flask
pymodm
dnspython
Pillow
matplotlib
pandas
scipy